@fabricior0cha [Task: Fixando issues #1]

+++ src/main/java/org/generation/blogPessoal/controller/PostagemController.java
@@ -1,6 +1,7 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -17,7 +18,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 /**
  * Controlador responsável pela tabela Postagem Com os métodos HTTP necessários
@@ -29,7 +32,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/postagens")
-@CrossOrigin("*")
+@CrossOrigin(allowedHeaders = "*", origins = "*")
 public class PostagemController {
 
 	@Autowired
@@ -60,13 +63,18 @@ public class PostagemController {
 		return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> delete(@PathVariable long id) {
-		return repository.findById(id).map(resp -> {
-			repository.deleteById(id);
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-		}).orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+	public void deletePostagem(@PathVariable Long id) {
 		
+		Optional<Postagem> postagem = repository.findById(id);
+		
+			if (postagem.isEmpty()) {
+				throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+			}
+		
+			repository.deleteById(id);	
+			
 	}
 
 }
