+++ src/main/java/org/generation/blogPessoal/controller/TemaController.java
@@ -1,6 +1,7 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -17,7 +18,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 /**
  * Controlador responsável pela tabela Tema Com os métodos HTTP necessários
@@ -60,12 +63,19 @@ public class TemaController {
 		return ResponseEntity.ok(repository.save(tema));
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> delete(@PathVariable long id) {
-		return repository.findById(id).map(resp -> {
+	public void deleteTema(@PathVariable Long id) {
+			
+		Optional<Tema> tema = repository.findById(id);
+		
+			if (tema.isEmpty()) {
+				
+				throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+			}
+			
 			repository.deleteById(id);
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-		}).orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+		
 	}
 
 }
